evaluators/azure/langevals_azure/content_safety.py
from typing import Literal
from azure.ai.contentsafety import ContentSafetyClient
from azure.core.credentials import AzureKeyCredential
from azure.ai.contentsafety.models import AnalyzeTextOptions

from langevals_core.base_evaluator import (
    BaseEvaluator,
    EvaluationResult,
    SingleEvaluationResult,
    EvaluatorParams,
)
from pydantic import BaseModel, Field


class AzureContentSafetyParams(EvaluatorParams):
    input: str


class AzureContentSafetySettings(BaseModel):
    severity_threshold: int = Field(ge=1, le=8, default=1)
    categories: list[str] = ["Hate", "SelfHarm", "Sexual", "Violence"]
    output_type: Literal["FourSeverityLevels", "EightSeverityLevels"] = (
        "FourSeverityLevels"
    )


class AzureContentSafetyEvaluator(
    BaseEvaluator[AzureContentSafetyParams, AzureContentSafetySettings]
):
    category = "safety"
    env_vars = ["AZURE_CONTENT_SAFETY_ENDPOINT", "AZURE_CONTENT_SAFETY_KEY"]

    def evaluate(self, params: AzureContentSafetyParams) -> SingleEvaluationResult:
        endpoint = self.env("AZURE_CONTENT_SAFETY_ENDPOINT")
        key = self.env("AZURE_CONTENT_SAFETY_KEY")

        client = ContentSafetyClient(endpoint, AzureKeyCredential(key))
        request = AnalyzeTextOptions(
            text=params.input,
            categories=self.settings.categories,
            output_type=self.settings.output_type,
        )
        response = client.analyze_text(request)

        categories_analysis = {
            item.category: item.severity for item in response.categories_analysis
        }
        score = max(categories_analysis.values(), default=0)  # type: ignore
        passed = score < self.settings.severity_threshold

        details = (
            ", ".join(
                f"{category} (severity {severity})"
                for category, severity in categories_analysis.items()
                if (severity or 0) >= self.settings.severity_threshold
            )
            or None
        )
        details = "Detected: " + details if details else None

        return EvaluationResult(score=score, passed=passed, details=details)
